// src/types/Location.ts
/**
 * A Location is a place that can be saved, deleted and viewed by the user.
 * This type represent a Location that already exists in the database, because it has a CUID generated at creation time.
 *
 * @typedef {Object} Location
 * @property {string} cuid - The CUID is a unique iden8tifier that is generated by the cuid package, it is usually a good
 * idea to use a public identifier for your documents instead of the _id field, it should be indexed for faster lookups
 * @property {string} name - The name of the location, e.g. "My house"
 * @property {number} latitude - The latitude of the location, e.g. 47.123456
 * @property {number} longitude - The longitude of the location, e.g. 23.123456
 * @property {string} address - The address of the location, e.g. "Some street, 123, Some city, Some country"
 */
export interface Location {
  cuid: string;
  name: string;
  latitude: number;
  longitude: number;
  address: string;
}